server/pyproject.toml
[project]
name = "voice-live-agent"
version = "0.1.0"
description = "Real-time voice agent with Azure Voice Live and ACS/Web streaming clients"
dependencies = [
    "aiohttp==3.11.18",
    "azure-core==1.34.0",
    "azure-identity==1.22.0",
    "certifi==2025.4.26",
    "cffi==1.17.1",
    "cryptography==44.0.3",
    "numpy==2.2.5",
    "pycparser==2.22",
    "python-dotenv==1.1.0",
    "requests==2.32.3",
    "sounddevice==0.5.1",
    "typing_extensions==4.13.2",
    "urllib3==2.4.0",
    "websocket-client==1.8.0",
    "quart",
    "websockets",
    "httpx",
    "quart-cors",
    "azure-storage-blob",
    "azure-cosmos"
]
requires-python = ">=3.9"
